CartItemList ausgelagert, weil es sowohl beim Overlay, als auch auf der seite angezeigt wird
