+++ src/pages/Cart.css
@@ -7,31 +7,6 @@
   box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
 }
 
-.cart-items-wrapper {
-  max-height: 300px;
-  overflow-y: auto;
-  border: 1px solid #ddd;
-  border-radius: 10px;
-  padding: 10px;
-  background: #f9f9f9;
-}
-
-.list-cartItems {
-  list-style-type: none;
-  padding: 0;
-  margin: 0;
-}
-
-.list-cartItems .cart-item:first-child {
-  border-top-left-radius: 10px;
-  border-top-right-radius: 10px;
-}
-
-.list-cartItems .cart-item:last-child {
-  border-bottom-left-radius: 10px;
-  border-bottom-right-radius: 10px;
-}
-
 .cart-total {
   margin-top: 15px;
   text-align: center;
